pyproject.toml
[project]
name = "llmtesthelper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi>=0.119.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "passlib>=1.7.4",
    "pdfplumber>=0.11.8",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "python-jose>=3.5.0",
    "python-json-logger>=4.0.0",
    "langchain>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pylint>=4.0.4",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
